backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
python-dotenv>=1.0.1
pydantic>=2.9.2
python-nmap>=0.7.1
slowapi>=0.1.9
httpx>=0.27.2
